requirements.txt
Flask==2.0.3
Flask-Bcrypt==0.7.1
pymongo==3.12.1
python-dotenv==0.20.0
bson==0.5.10  # Only needed if you're not using pymongo
google-generativeai  # You might need to specify a version if available
